requirement.txt
Django==2.2.19
djangorestframework==3.11.0
PyJWT==2.6.0